Move toContain to its own file
