+++ src/Corretto/Expect.php
@@ -45,8 +45,9 @@ class Expectation {
 	}
 
 	public function toEqual( $expected ) {
-		$actual = $this->recursiveAssocSort( $this->actual );
-		$expected = $this->recursiveAssocSort( $expected );
+		$helper = new Helpers();
+		$actual = $helper->recursiveAssocSort( $this->actual );
+		$expected = $helper->recursiveAssocSort( $expected );
 		if ( $expected !== $actual ) {
 			$expectedString = var_export( $expected, true );
 			$actualString = var_export( $actual, true );
@@ -94,59 +95,6 @@ class Expectation {
 			throw new AssertionFailure( "Failed asserting that " . $actualString . " is less than " . $expectedString . "" );
 		}
 	}
-
-	private function recursiveAssocSort( $elements ) {
-		if ( ! is_array( $elements ) ) {
-			return $elements;
-		}
-		$elements = array_map( function( $el ) {
-			return $this->recursiveAssocSort( $el );
-		}, $elements );
-		if ( ! $this->hasStringKeys( $elements ) ) {
-			return $elements;
-		}
-		sort( $elements );
-		return $elements;
-	}
-
-	private function hasStringKeys( array $array ) {
-		return count( array_filter( array_keys( $array ), 'is_string' ) ) > 0;
-	}
-}
-
-class ContainExpectation {
-	function __construct( $actual ) {
-		$this->actual = $actual;
-	}
-
-	public function toContain( $expected ) {
-		$actual = $this->actual;
-		if ( is_string( $actual ) && strpos( $actual, $expected ) === false ) {
-			$expectedString = var_export( $expected, true );
-			$actualString = var_export( $actual, true );
-			throw new AssertionFailure( "Failed asserting that " . $actualString . " contains " . $expectedString . "" );
-		}
-		if ( is_array( $actual ) && ! in_array( $expected, $actual ) ) {
-			$expectedString = var_export( $expected, true );
-			$actualString = var_export( $actual, true );
-			throw new AssertionFailure( "Failed asserting that " . $actualString . " contains " . $expectedString . "" );
-		}
-	}
-
-	public function toNotContain( $expected ) {
-		$actual = $this->actual;
-		if ( is_string( $actual ) && strpos( $actual, $expected ) !== false ) {
-			$expectedString = var_export( $expected, true );
-			$actualString = var_export( $actual, true );
-			throw new AssertionFailure( "Failed asserting that " . $actualString . " does not contain " . $expectedString . "" );
-		}
-		if ( is_array( $actual ) && in_array( $expected, $actual ) ) {
-			$expectedString = var_export( $expected, true );
-			$actualString = var_export( $actual, true );
-			throw new AssertionFailure( "Failed asserting that " . $actualString . " does not contain " . $expectedString . "" );
-		}
-	}
 }
 
 extendExpectation( '\Corretto\ContainExpectation' );
-
